+++ packages/terra-notification-dialog/src/terra-dev-site/doc/example/NotificationDialogWithFocus.jsx
@@ -54,12 +54,12 @@ class NotificationDialogWithFocus extends React.Component {
           acceptAction={{
             text: 'Confirm',
             onClick: clickConfirm,
-            isEmphasized: true,
           }}
           rejectAction={{
             text: 'Close',
             onClick: this.handleCloseModal,
           }}
+          emphasizedAction="accept"
         />
         <Button text="Trigger NotificationDialog" onClick={this.handleOpenModal} />
         <Button text="Dismiss" onClick={this.props.disclosureManager.dismiss} />