+++ packages/terra-notification-dialog/src/terra-dev-site/doc/example/NotificationDialogWithLongText.jsx
@@ -116,12 +116,12 @@ class NotificationDialogWithLongText extends React.Component {
           acceptAction={{
             text: 'Confirm',
             onClick: clickConfirm,
-            isEmphasized: true,
           }}
           rejectAction={{
             text: 'Close',
             onClick: this.handleCloseModal,
           }}
+          emphasizedAction="accept"
         />
         <Button text="Trigger NotificationDialog" onClick={this.handleOpenModal} />
       </div>