+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/ContentNotificationDialog.test.jsx
@@ -44,12 +44,12 @@ class ContentNotificationDialog extends React.Component {
           acceptAction={{
             text: 'Confirm',
             onClick: clickConfirm,
-            isEmphasized: true,
           }}
           rejectAction={{
             text: 'Close',
             onClick: this.handleCloseModal,
           }}
+          emphasizedAction="accept"
         />
         <Button id="trigger-notification-dialog" text="Trigger NotificationDialog" onClick={this.handleOpenModal} />
       </div>