+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/EmphasizedRejectNotificationDialog.test.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import Button from 'terra-button';
-import NotificationDialog, { NotificationDialogVariants } from 'terra-notification-dialog';
+import NotificationDialog, { NotificationDialogVariants } from '../../../NotificationDialog';
 
 const clickConfirm = () => {
-  console.log('You clicked confirm'); // eslint-disable-line no-console
+  alert('You clicked confirm'); // eslint-disable-line no-alert
 };
 
-class CompleteNotificationDialog extends React.Component {
+class ReversedActionNotificationDialog extends React.Component {
   constructor() {
     super();
 
@@ -42,12 +42,12 @@ class CompleteNotificationDialog extends React.Component {
             text: 'Close',
             onClick: this.handleCloseModal,
           }}
-          emphasizedAction="accept"
+          emphasizedAction="reject"
         />
-        <Button text="Trigger NotificationDialog" onClick={this.handleOpenModal} />
+        <Button id="trigger-notification-dialog" text="Trigger NotificationDialog" onClick={this.handleOpenModal} />
       </div>
     );
   }
 }
 
-export default CompleteNotificationDialog;
+export default ReversedActionNotificationDialog;