+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/NoVariantNotificationDialog.test.jsx
@@ -37,12 +37,12 @@ class NoVariantNotificationDialog extends React.Component {
           acceptAction={{
             text: 'Confirm',
             onClick: clickConfirm,
-            isEmphasized: true,
           }}
           rejectAction={{
             text: 'Close',
             onClick: this.handleCloseModal,
           }}
+          emphasizedAction="accept"
         />
         <Button id="trigger-notification-dialog" text="Trigger Notification Dialog" onClick={this.handleOpenModal} />
       </div>