+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/NotificationDialogMinimalProps.test.jsx
@@ -30,8 +30,8 @@ class NotificationDialogMinimalProps extends React.Component {
           acceptAction={{
             text: 'Close',
             onClick: this.handleCloseModal,
-            isEmphasized: true,
           }}
+          emphasizedAction="accept"
         />
         <Button id="trigger-notification-dialog" text="Trigger NotificationDialog" onClick={this.handleOpenModal} />
       </div>