+++ packages/terra-date-picker/src/terra-dev-site/test/date-picker/DatePickerIgnoreInvalidSelectedDates.test.jsx
@@ -1,8 +1,12 @@
 import React from 'react';
+import classNames from 'classnames/bind';
 import DatePicker from '../../../DatePicker';
+import styles from './common/DatePicker.test.module.scss';
+
+const cx = classNames.bind(styles);
 
 const DateTimePickerInvalidSelectedDates = () => (
-  <div>
+  <div className={cx('content-wrapper')}>
     <span id="selected-date">selectedDate = null</span>
     <br />
     <DatePicker