+++ packages/terra-date-picker/src/terra-dev-site/test/date-picker/DatePickerSelectedDate.test.jsx
@@ -1,6 +1,10 @@
 import React from 'react';
 import moment from 'moment';
+import classNames from 'classnames/bind';
 import DatePicker from '../../../DatePicker';
+import styles from './common/DatePicker.test.module.scss';
+
+const cx = classNames.bind(styles);
 
 class DatePickerDefault extends React.Component {
   constructor(props) {
@@ -17,7 +21,7 @@ class DatePickerDefault extends React.Component {
 
   render() {
     return (
-      <div>
+      <div className={cx('content-wrapper')}>
         <DatePicker
           name="date-input"
           selectedDate={this.state.date}