+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -9,7 +9,7 @@ Unreleased
 * Renamed 'secondaryAction' prop to 'rejectAction'
 
 ### Added
-* Added an additional 'isEmphasized' bool to the 'acceptAction' prop object shape to allow changing the resultant button to non-emphasized.
+* Added a 'emphasizedAction' oneOf to choose between emphasizing the accept button, reject button, or neither.
 * Added a 'content' node prop that will display the given node in the Dialog after 'startMessage'
 * Added an 'endMessage' string prop that will display a message after both 'startMessage' and 'content'