+++ packages/terra-notification-dialog/CHANGELOG.md
@@ -4,6 +4,22 @@ Changelog
 Unreleased
 ----------
 
+3.15.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+### Fixed
+* Added terra-breakpoints to dependencies for the project
+
+3.14.0 - (October 3, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+### Fixed
+* Removed extra comma from scss file.
+
 3.13.0 - (September 26, 2019)
 ------------------
 ### Changed